requirements.txt
accelerate==1.2.1
huggingface_hub==0.26.3
transformers==4.46.3
diffusers==0.32.2
peft==0.13.2
sentencepiece==0.2.0
flash-attn==2.7.3
einops==0.8.1
cupy-cuda12x==13.5.1
numpy==1.26.4
torch_scatter==2.1.2
spconv-cu121==2.3.8
timm==1.0.19
openexr==3.3.5
trimesh==4.7.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
opencv-python==4.7.0.72
omegaconf==2.3.0
addict==2.4.0
protobuf==6.31.1
ninja==1.11.1.4
git+https://github.com/NVlabs/nvdiffrast
git+https://github.com/mit-han-lab/torchsparse.git